Introduce website getKeepStream endpoint
