+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/mobile/MobileRecommendationsController.scala
@@ -122,15 +122,6 @@ class MobileRecommendationsController @Inject() (
     }
   }
 
-  def keepUpdates(limit: Int, beforeId: Option[String], afterId: Option[String]) = UserAction.async { request =>
-    val beforeExtId = beforeId.flatMap(id => ExternalId.asOpt[Keep](id))
-    val afterExtId = afterId.flatMap(id => ExternalId.asOpt[Keep](id))
-
-    commander.updatesFromFollowedLibraries(request.userId, limit, beforeExtId, afterExtId).map { updatedKeeps =>
-      Ok(Json.obj("updatedKeeps" -> updatedKeeps))
-    }
-  }
-
   private def getRecommendationSource(request: UserRequest[_]): RecommendationSource = {
     val agent = UserAgent(request)
     if (agent.isKifiAndroidApp) RecommendationSource.Android
