+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/website/KeepsController.scala
@@ -236,4 +236,13 @@ class KeepsController @Inject() (
       Ok(Json.obj("results" -> results))
     }
   }
+
+  def getKeepStream(limit: Int, beforeId: Option[String], afterId: Option[String]) = UserAction.async { request =>
+    val beforeExtId = beforeId.flatMap(id => ExternalId.asOpt[Keep](id))
+    val afterExtId = afterId.flatMap(id => ExternalId.asOpt[Keep](id))
+
+    keepsCommander.getKeepStream(request.userId, limit, beforeExtId, afterExtId).map { keeps =>
+      Ok(Json.obj("keeps" -> keeps))
+    }
+  }
 }
